+++ common/src/main/java/us/myles/ViaVersion/protocols/protocol1_16_4_5to1_16_3/Protocol1_16_4_5To1_16_3.java
@@ -1,4 +1,4 @@
-package us.myles.ViaVersion.protocols.protocol1_16_4to1_16_3;
+package us.myles.ViaVersion.protocols.protocol1_16_4_5to1_16_3;
 
 import us.myles.ViaVersion.api.protocol.Protocol;
 import us.myles.ViaVersion.api.remapper.PacketRemapper;
@@ -6,9 +6,9 @@ import us.myles.ViaVersion.api.type.Type;
 import us.myles.ViaVersion.protocols.protocol1_16_2to1_16_1.ClientboundPackets1_16_2;
 import us.myles.ViaVersion.protocols.protocol1_16_2to1_16_1.ServerboundPackets1_16_2;
 
-public class Protocol1_16_4To1_16_3 extends Protocol<ClientboundPackets1_16_2, ClientboundPackets1_16_2, ServerboundPackets1_16_2, ServerboundPackets1_16_2> {
+public class Protocol1_16_4_5To1_16_3 extends Protocol<ClientboundPackets1_16_2, ClientboundPackets1_16_2, ServerboundPackets1_16_2, ServerboundPackets1_16_2> {
 
-    public Protocol1_16_4To1_16_3() {
+    public Protocol1_16_4_5To1_16_3() {
         super(ClientboundPackets1_16_2.class, ClientboundPackets1_16_2.class, ServerboundPackets1_16_2.class, ServerboundPackets1_16_2.class);
     }
