+++ common/src/main/java/us/myles/ViaVersion/protocols/protocol1_17to1_16_4_5/storage/EntityTracker1_17.java
@@ -1,4 +1,4 @@
-package us.myles.ViaVersion.protocols.protocol1_17to1_16_4.storage;
+package us.myles.ViaVersion.protocols.protocol1_17to1_16_4_5.storage;
 
 import us.myles.ViaVersion.api.data.UserConnection;
 import us.myles.ViaVersion.api.entities.Entity1_17Types.EntityType;
