+++ common/src/main/java/us/myles/ViaVersion/protocols/protocol1_17to1_16_4_5/types/Chunk1_17Type.java
@@ -1,4 +1,4 @@
-package us.myles.ViaVersion.protocols.protocol1_17to1_16_4.types;
+package us.myles.ViaVersion.protocols.protocol1_17to1_16_4_5.types;
 
 import com.github.steveice10.opennbt.tag.builtin.CompoundTag;
 import com.google.common.base.Preconditions;
